requirements.txt
pyautogui==0.9.52
pafy==0.5.5
pyinstaller==4.0
python-twitter==3.5
python-vlc==3.0.11115
requests==2.24.0
youtube-dl==2020.9.6